requirements.txt
astroid==2.4.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.3
Pillow==7.1.2
pylint==2.5.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
wrapt==1.12.1